tech: fixing the postgresql
