+++ src/main/resources/cts/schema/postgresql/v1_5/v1_5_0__changes_postgresql.sql
@@ -6,6 +6,8 @@ CREATE TABLE IF NOT EXISTS cts_active_nodes (
   CTSAN_LAST_SEEN   TIMESTAMP               NOT NULL
 );
 
+ALTER TABLE CLUSTER_TASK_META DROP COLUMN CTSKM_MAX_TIME_TO_RUN;
+
 CREATE OR REPLACE FUNCTION insert_task(
   task_type INTEGER,
   processor_type CHARACTER VARYING(40),
