+++ src/test/resources/db.properties
@@ -1,10 +1,10 @@
 #
 #   MSSQL (local)
 #
-type=MSSQL
-url=jdbc:sqlserver://localhost:1433;databaseName=cts_tests_schema
-username=sa
-password=Mercurypw123
+#type=MSSQL
+#url=jdbc:sqlserver://localhost:1433;databaseName=cts_tests_schema
+#username=sa
+#password=Mercurypw123
 #
 #   MSSQL (remote) (-Dtests.db.url=jdbc:sqlserver://mydwbld0150.hpeswlab.net:1433 -Dtests.db.username=sa -Dtests.db.password=Mercurypw123)
 #
@@ -22,7 +22,7 @@ password=Mercurypw123
 #
 #   POSTGRESQL (local)
 #
-#type=POSTGRESQL
-#url=jdbc:postgresql://localhost:5432/cts_tests_schema
-#username=postgres
-#password=postgres
+type=POSTGRESQL
+url=jdbc:postgresql://localhost:5432/cts_tests_schema
+username=postgres
+password=postgres
